cis_Jiekou/cisPytest/requirements.txt
pytest
pytest-htm;
pytest-xdist
pytest-ordering
pytest-rerunfailres
allure-pytest